# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Hevy API Docs",
        "description": "\n      Welcome to Hevy's public API! We're just starting to roll this out and depending on your\n      feedback, we'll be adding more features and endpoints. Also, we make no guarantees that \n      we won't completely change the structure or abandon the project entirely so use it at\n      your own risk. Currently, this API is only available to Hevy Pro users. You can get your \n      key on our web app at https://hevy.com/settings?developer. If you have any questions,\n      please reach out to pedro@hevyapp.com 🤝.\n      ",
        "version": "0.0.1"
    },
    "paths": {
...
        "/v1/routines": {
            "get": {
                "summary": "Get a paginated list of routines",
                "tags": [
                    "Routines"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "api-key",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        },
                        "description": "Page number (Must be 1 or greater)"
                    },
                    {
                        "in": "query",
                        "name": "pageSize",
                        "schema": {
                            "type": "integer",
                            "default": 5
                        },
                        "description": "Number of items on the requested page (Max 10)"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A paginated list of routines",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "page": {
                                            "type": "integer",
                                            "example": 1,
                                            "description": "Current page number"
                                        },
                                        "page_count": {
                                            "type": "integer",
                                            "example": 5,
                                            "description": "Total number of pages"
                                        },
                                        "routines": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Routine"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid page size"
                    }
                }
            },
            "post": {
                "summary": "Create a new routine",
                "tags": [
                    "Routines"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "api-key",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostRoutinesRequestBody"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "The routine was successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Routine"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "description": "Error message"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Routine limit exceeded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "description": "Error message"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/routines/{routineId}": {
            "get": {
                "summary": "Get a routine by its Id",
                "tags": [
                    "Routines"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "api-key",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true
                    },
                    {
                        "in": "path",
                        "name": "routineId",
                        "description": "The id of the routine",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The routine with the provided id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "routine": {
                                            "$ref": "#/components/schemas/Routine"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "description": "Error message"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update an existing routine",
                "tags": [
                    "Routines"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "api-key",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true
                    },
                    {
                        "in": "path",
                        "name": "routineId",
                        "description": "The id of the routine",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PutRoutinesRequestBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The routine was successfully updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Routine"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "description": "Error message"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Routine doesn't exist or doesn't belong to the user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "description": "Error message"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/routine_folders": {
            "get": {
                "summary": "Get a paginated list of routine folders available on the account.",
                "tags": [
                    "RoutineFolders"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "api-key",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        },
                        "description": "Page number (Must be 1 or greater)"
                    },
                    {
                        "in": "query",
                        "name": "pageSize",
                        "schema": {
                            "type": "integer",
                            "default": 5
                        },
                        "description": "Number of items on the requested page (Max 10)"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A paginated list of routine folders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "page": {
                                            "type": "integer",
                                            "default": 1,
                                            "description": "Current page number"
                                        },
                                        "page_count": {
                                            "type": "integer",
                                            "default": 5,
                                            "description": "Total number of pages"
                                        },
                                        "routine_folders": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/RoutineFolder"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid page size"
                    }
                }
            },
            "post": {
                "summary": "Create a new routine folder. The folder will be created at index 0, and all other folders will have their indexes incremented.",
                "tags": [
                    "RoutineFolders"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "api-key",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostRoutineFolderRequestBody"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "The routine folder was successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoutineFolder"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "description": "Error message"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/routine_folders/{folderId}": {
            "get": {
                "tags": [
                    "RoutineFolders"
                ],
                "summary": "Get a single routine folder by id.",
                "parameters": [
                    {
                        "in": "header",
                        "name": "api-key",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true
                    },
                    {
                        "name": "folderId",
                        "in": "path",
                        "description": "The id of the routine folder",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoutineFolder"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Routine folder not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PostRoutinesRequestExercise": {
                "type": "object",
                "properties": {
                    "exercise_template_id": {
                        "type": "string",
                        "description": "The ID of the exercise template.",
                        "example": "D04AC939"
                    },
                    "superset_id": {
                        "type": "integer",
                        "nullable": true,
                        "description": "The ID of the superset.",
                        "example": null
                    },
                    "rest_seconds": {
                        "type": "integer",
                        "nullable": true,
                        "description": "The rest time in seconds.",
                        "example": 90
                    },
                    "notes": {
                        "type": "string",
                        "nullable": true,
                        "description": "Additional notes for the exercise.",
                        "example": "Stay slow and controlled."
                    },
                    "sets": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PostRoutinesRequestSet"
                        }
                    }
                }
            },
            "PostRoutinesRequestBody": {
                "type": "object",
                "properties": {
                    "routine": {
                        "type": "object",
                        "properties": {
                            "title": {
                                "type": "string",
                                "description": "The title of the routine.",
                                "example": "April Leg Day 🔥"
                            },
                            "folder_id": {
                                "type": "number",
                                "nullable": true,
                                "description": "The folder id the routine should be added to. Pass null to insert the routine into default \"My Routines\" folder",
                                "example": null
                            },
                            "notes": {
                                "type": "string",
                                "description": "Additional notes for the routine.",
                                "example": "Focus on form over weight. Remember to stretch."
                            },
                            "exercises": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/PostRoutinesRequestExercise"
                                }
                            }
                        }
                    }
                }
            },
...
            "PutRoutinesRequestBody": {
                "type": "object",
                "properties": {
                    "routine": {
                        "type": "object",
                        "properties": {
                            "title": {
                                "type": "string",
                                "description": "The title of the routine.",
                                "example": "April Leg Day 🔥"
                            },
                            "notes": {
                                "type": "string",
                                "nullable": true,
                                "description": "Additional notes for the routine.",
                                "example": "Focus on form over weight. Remember to stretch."
                            },
                            "exercises": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/PutRoutinesRequestExercise"
                                }
                            }
                        }
                    }
                }
            },
            "PostRoutineFolderRequestBody": {
                "type": "object",
                "properties": {
                    "routine_folder": {
                        "type": "object",
                        "properties": {
                            "title": {
                                "type": "string",
                                "description": "The title of the routine folder.",
                                "example": "Push Pull 🏋️‍♂️"
                            }
                        }
                    }
                }
            },
...
            "RoutineFolder": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number",
                        "description": "The routine folder ID.",
                        "example": 42
                    },
                    "index": {
                        "type": "number",
                        "description": "The routine folder index. Describes the order of the folder in the list.",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "description": "The routine folder title.",
                        "example": "Push Pull 🏋️‍♂️"
                    },
                    "updated_at": {
                        "type": "string",
                        "description": "ISO 8601 timestamp of when the folder was last updated.",
                        "example": "2021-09-14T12:00:00Z"
                    },
                    "created_at": {
                        "type": "string",
                        "description": "ISO 8601 timestamp of when the folder was created.",
                        "example": "2021-09-14T12:00:00Z"
                    }
                }
            },
            "Routine": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "The routine ID.",
                        "example": "b459cba5-cd6d-463c-abd6-54f8eafcadcb"
                    },
                    "title": {
                        "type": "string",
                        "description": "The routine title.",
                        "example": "Upper Body 💪"
                    },
                    "folder_id": {
                        "type": "number",
                        "nullable": true,
                        "description": "The routine folder ID.",
                        "example": 42
                    },
                    "updated_at": {
                        "type": "string",
                        "description": "ISO 8601 timestamp of when the routine was last updated.",
                        "example": "2021-09-14T12:00:00Z"
                    },
                    "created_at": {
                        "type": "string",
                        "description": "ISO 8601 timestamp of when the routine was created.",
                        "example": "2021-09-14T12:00:00Z"
                    },
                    "exercises": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "index": {
                                    "type": "number",
                                    "description": "Index indicating the order of the exercise in the routine.",
                                    "example": 0
                                },
                                "title": {
                                    "type": "string",
                                    "description": "Title of the exercise",
                                    "example": "Bench Press (Barbell)"
                                },
                                "rest_seconds": {
                                    "type": "string",
                                    "description": "The rest time in seconds between sets of the exercise",
                                    "example": 60
                                },
                                "notes": {
                                    "type": "string",
                                    "description": "Routine notes on the exercise",
                                    "example": "Focus on form. Go down to 90 degrees."
                                },
                                "exercise_template_id": {
                                    "type": "string",
                                    "description": "The id of the exercise template. This can be used to fetch the exercise template.",
                                    "example": "05293BCA"
                                },
                                "supersets_id": {
                                    "type": "number",
                                    "nullable": true,
                                    "description": "The id of the superset that the exercise belongs to. A value of null indicates the exercise is not part of a superset.",
                                    "example": 0
                                },
                                "sets": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "index": {
                                                "type": "number",
                                                "description": "Index indicating the order of the set in the routine.",
                                                "example": 0
                                            },
                                            "type": {
                                                "type": "string",
                                                "description": "The type of set. This can be one of 'normal', 'warmup', 'dropset', 'failure'",
                                                "example": "normal"
                                            },
                                            "weight_kg": {
                                                "type": "number",
                                                "nullable": true,
                                                "description": "Weight lifted in kilograms.",
                                                "example": 100
                                            },
                                            "reps": {
                                                "type": "number",
                                                "nullable": true,
                                                "description": "Number of reps logged for the set",
                                                "example": 10
                                            },
                                            "rep_range": {
                                                "type": "object",
                                                "nullable": true,
                                                "description": "Range of reps for the set, if applicable",
                                                "properties": {
                                                    "start": {
                                                        "type": "number",
                                                        "nullable": true,
                                                        "description": "Starting rep count for the range",
                                                        "example": 8
                                                    },
                                                    "end": {
                                                        "type": "number",
                                                        "nullable": true,
                                                        "description": "Ending rep count for the range",
                                                        "example": 12
                                                    }
                                                }
                                            },
                                            "distance_meters": {
                                                "type": "number",
                                                "nullable": true,
                                                "description": "Number of meters logged for the set",
                                                "example": null
                                            },
                                            "duration_seconds": {
                                                "type": "number",
                                                "nullable": true,
                                                "description": "Number of seconds logged for the set",
                                                "example": null
                                            },
                                            "rpe": {
                                                "type": "number",
                                                "nullable": true,
                                                "description": "RPE (Relative perceived exertion) value logged for the set",
                                                "example": 9.5
                                            },
                                            "custom_metric": {
                                                "type": "number",
                                                "nullable": true,
                                                "description": "Custom metric logged for the set (Currently only used to log floors or steps for stair machine exercises)",
                                                "example": 50
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
...
        }
    },
...
}